refactor(haku): name tool rendering vocabulary
