+++ haku/console/frontend/tool_rendering/vocabulary.tsx
@@ -1,4 +1,4 @@
-// Compact vs detailed rendering, shared by the per-server preview widgets. A **compact**
+// Shared vocabulary for per-server preview widgets, including compact vs detailed rendering. A **compact**
 // preview is the scannable form shown on a drawer approval card (and anywhere space is
 // tight): list-shaped arguments collapse to their first few items, long bodies to their
 // first few lines. A **detailed** preview is the full form shown in the expanded detail
